Remove training.runtime package
